NAVIO-1501: Making Example Model Tests work in Jenkins (#27)
* feat(NAVIO-1501): fixing ci job hanging issue

* feat(NAVIO-1501): fixing ci job hanging issue

* feat(NAVIO-1501): adding example tests and coverage

Also moved Makefile and README.md in examples

* fix(NAVIO-1800): mock pump leakage data instead of download (#29)

* NAVIO-1748: Demo Model for allowed Data Types (#26)

* feat(NAVIO-1748): demo model for allowed data types

* feat(NAVIO-1748): type demo model test

* feat(NAVIO-1748): correctly handling integer fields

See also NAVIO-1804

* feat(NAVIO-1748): fix tests

* feat(NAVIO-1501): fixing dependency file-only test

* feat(NAVIO-1639): visual inspection example model (#28)

* NAVIO-1845: Moving Example Models From Predictor Pipeline Repo (#30)

* feat(NAVIO-1845): moving example models from predictor-pipeline

Also fixing:

- NAVIO-1873
- NAVIO-1874

* feat(NAVIO-1845): docstrings for traits

+++ tests/test_pynavio/test_schema.py
@@ -1,10 +1,10 @@
 import unittest
 from typing import Any, Dict, List
+
 import pandas as pd
 import pytest
 
 from pynavio import make_example_request
-
 
 DATA = {'x': 1, 'y': 2., 'z': 'str', 't': '2020-01-01 00:00:00'}
 
@@ -12,7 +12,7 @@ DATA_FRAME = pd.DataFrame([DATA])
 
 SCHEMA = {
     'x': {
-        'type': 'float',
+        'type': 'int',
         'nullable': False
     },
     'y': {
